# Conflicts:
#	src-ui-wx/app-shell/KeyBindingEditDialog.cpp
#	src-ui-wx/preferences/OtherSettingsPanel.cpp

+++ src-core/utils/VectorMath.cpp
@@ -248,7 +248,7 @@ glm::vec2 VectorMath::GetScreenCoord(
     glm::vec4 clipSpacePos = MVP * glm::vec4(position, 1.0);
     if (clipSpacePos.w == 0.0f) return glm::vec2(0.0f);
     glm::vec3 ndcSpacePos = glm::vec3(clipSpacePos.x / clipSpacePos.w, clipSpacePos.y / clipSpacePos.w, clipSpacePos.z / clipSpacePos.w);
-    glm::vec2 screen_position(((ndcSpacePos.x + 1.0) / 2.0) * screenWidth, ((1.0 - ndcSpacePos.y) / 2.0) * screenHeight);
+    glm::vec2 screen_position(((ndcSpacePos.x + 1.0) / 2.0) * ((float)screenWidth), ((1.0 - ndcSpacePos.y) / 2.0) * ((float)screenHeight));
     return screen_position;
 }
